+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/assets/application/base/service/AssetBaseService.java
@@ -30,6 +30,7 @@ import java.util.List;
 import java.util.Map;
 
 public interface AssetBaseService {
+
     void synchronizeAssetsAsync(List<String> globalAssetIds);
 
     void synchronizeAssetsAsync(String globalAssetId);