+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/assets/infrastructure/base/irs/IrsService.java
@@ -22,40 +22,57 @@
 package org.eclipse.tractusx.traceability.assets.infrastructure.base.irs;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.eclipse.tractusx.irs.edc.client.policy.OperatorType;
 import org.eclipse.tractusx.traceability.assets.domain.base.IrsRepository;
 import org.eclipse.tractusx.traceability.assets.domain.base.model.AssetBase;
+import org.eclipse.tractusx.traceability.assets.domain.base.model.Owner;
 import org.eclipse.tractusx.traceability.assets.infrastructure.base.irs.model.request.BomLifecycle;
 import org.eclipse.tractusx.traceability.assets.infrastructure.base.irs.model.request.RegisterJobRequest;
 import org.eclipse.tractusx.traceability.assets.infrastructure.base.irs.model.request.RegisterPolicyRequest;
 import org.eclipse.tractusx.traceability.assets.infrastructure.base.irs.model.response.Direction;
 import org.eclipse.tractusx.traceability.assets.infrastructure.base.irs.model.response.JobDetailResponse;
-import org.eclipse.tractusx.traceability.assets.infrastructure.base.irs.model.response.JobStatus;
 import org.eclipse.tractusx.traceability.assets.infrastructure.base.irs.model.response.PolicyResponse;
-import org.eclipse.tractusx.traceability.assets.infrastructure.base.irs.model.response.RegisterJobResponse;
 import org.eclipse.tractusx.traceability.bpn.domain.service.BpnRepository;
 import org.eclipse.tractusx.traceability.common.properties.TraceabilityProperties;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 
-import java.util.Collections;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Slf4j
 @Service
-@RequiredArgsConstructor
 public class IrsService implements IrsRepository {
 
     private final IRSApiClient irsClient;
     private final BpnRepository bpnRepository;
     private final TraceabilityProperties traceabilityProperties;
     private final ObjectMapper objectMapper;
+    private final AssetCallbackRepository assetAsBuiltCallbackRepository;
+    private final AssetCallbackRepository assetAsPlannedCallbackRepository;
+
+    public IrsService(
+            IRSApiClient irsClient,
+            BpnRepository bpnRepository,
+            TraceabilityProperties traceabilityProperties,
+            ObjectMapper objectMapper,
+            @Qualifier("assetAsBuiltRepositoryImpl")
+            AssetCallbackRepository assetAsBuiltCallbackRepository,
+            @Qualifier("assetAsPlannedRepositoryImpl")
+            AssetCallbackRepository assetAsPlannedCallbackRepository) {
+        this.irsClient = irsClient;
+        this.bpnRepository = bpnRepository;
+        this.traceabilityProperties = traceabilityProperties;
+        this.objectMapper = objectMapper;
+        this.assetAsBuiltCallbackRepository = assetAsBuiltCallbackRepository;
+        this.assetAsPlannedCallbackRepository = assetAsPlannedCallbackRepository;
+    }
 
     @Override
-    public List<AssetBase> findAssets(String globalAssetId, Direction direction, List<String> aspects, BomLifecycle bomLifecycle) {
-        RegisterJobRequest registerJobRequest = RegisterJobRequest.buildJobRequest(globalAssetId, traceabilityProperties.getBpn().toString(), direction, aspects, bomLifecycle);
+    public void createJobToResolveAssets(String globalAssetId, Direction direction, List<String> aspects, BomLifecycle bomLifecycle) {
+        RegisterJobRequest registerJobRequest = RegisterJobRequest.buildJobRequest(globalAssetId, traceabilityProperties.getBpn().toString(), direction, aspects, bomLifecycle, traceabilityProperties.getUrl());
         log.info("Build HTTP Request {}", registerJobRequest);
         try {
             log.info("Build HTTP Request as JSON {}", objectMapper.writeValueAsString(registerJobRequest));
@@ -63,12 +80,18 @@ public class IrsService implements IrsRepository {
             log.error("exception", e);
         }
 
-        RegisterJobResponse startJobResponse = irsClient.registerJob(registerJobRequest);
-        JobDetailResponse jobResponse = irsClient.getJobDetails(startJobResponse.id());
+        irsClient.registerJob(registerJobRequest);
+    }
 
-        JobStatus jobStatus = jobResponse.jobStatus();
-        long runtime = (jobStatus.lastModifiedOn().getTime() - jobStatus.startedOn().getTime()) / 1000;
-        log.info("IRS call for globalAssetId: {} finished with status: {}, runtime {} s.", globalAssetId, jobStatus.state(), runtime);
+    @Override
+    public void handleJobFinishedCallback(String jobId, String state) {
+        if (!Objects.equals(state, JobDetailResponse.JOB_STATUS_COMPLETED)) {
+            return;
+        }
+        JobDetailResponse jobResponse = irsClient.getJobDetails(jobId);
+
+        long runtime = (jobResponse.jobStatus().lastModifiedOn().getTime() - jobResponse.jobStatus().startedOn().getTime()) / 1000;
+        log.info("IRS call for globalAssetId: {} finished with status: {}, runtime {} s.", jobResponse.jobStatus().globalAssetId(), jobResponse.jobStatus().state(), runtime);
         try {
             log.info("Received HTTP Response: {}", objectMapper.writeValueAsString(jobResponse));
         } catch (Exception e) {
@@ -81,9 +104,32 @@ public class IrsService implements IrsRepository {
             } catch (Exception e) {
                 log.warn("BPN Mapping Exception", e);
             }
-            return jobResponse.convertAssets();
+
+            // persist converted assets
+            jobResponse.convertAssets().forEach(assetBase -> {
+                if (assetBase.getBomLifecycle() == org.eclipse.tractusx.irs.component.enums.BomLifecycle.AS_BUILT) {
+                    saveOrUpdateAssets(assetAsBuiltCallbackRepository, assetBase);
+                } else if (assetBase.getBomLifecycle() == org.eclipse.tractusx.irs.component.enums.BomLifecycle.AS_PLANNED) {
+                    saveOrUpdateAssets(assetAsPlannedCallbackRepository, assetBase);
+                }
+            });
         }
-        return Collections.emptyList();
+    }
+
+    void saveOrUpdateAssets(AssetCallbackRepository repository, AssetBase asset) {
+        Optional<AssetBase> existingAssetOptional = repository.findById(asset.getId());
+        if (existingAssetOptional.isPresent()) {
+            AssetBase existingAsset = existingAssetOptional.get();
+            if (existingAsset.getOwner().equals(Owner.UNKNOWN)) {
+                existingAsset.setOwner(asset.getOwner());
+            }
+            if (!asset.getParentRelations().isEmpty()) {
+                existingAsset.setParentRelations(asset.getParentRelations());
+            }
+            repository.save(existingAsset);
+        } else {
+            repository.save(asset);
+        }
     }
 
     @Override
@@ -103,7 +149,7 @@ public class IrsService implements IrsRepository {
 
         //create missing policies
         boolean missingPolicy = irsPolicies.stream().noneMatch(irsPolicy -> irsPolicy.policyId().equals(traceabilityProperties.getRightOperand()));
-        if(missingPolicy){
+        if (missingPolicy) {
             createPolicy();
         }
     }
@@ -124,7 +170,7 @@ public class IrsService implements IrsRepository {
         }
     }
 
-    public List<PolicyResponse> getPolicies(){
+    public List<PolicyResponse> getPolicies() {
         return irsClient.getPolicies();
     }