+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/common/config/SecurityConfig.java
@@ -58,7 +58,8 @@ public class SecurityConfig {
             "/internal/endpoint-data-reference",
             "/actuator/**",
             "/registry/reload",
-            "/submodel/**"
+            "/submodel/**",
+            "/irs/job/callback"
 
     };