+++ tx-backend/src/test/java/org/eclipse/tractusx/traceability/integration/assets/AssetAsBuiltControllerSyncIT.java
@@ -26,7 +26,6 @@ import org.jose4j.lang.JoseException;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.annotation.DirtiesContext;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Map;
@@ -44,13 +43,12 @@ class AssetAsBuiltControllerSyncIT extends IntegrationTestSpecification {
     AssetsSupport assetsSupport;
 
     @Test
-    void shouldSynchronizeAssets() throws JoseException, InterruptedException {
-        //GIVEN
+    void givenSyncRequestWithOneAsset_whenSyncTriggered_thenShouldIrsApiOnve() throws JoseException, InterruptedException {
+        // GIVEN
         oAuth2ApiSupport.oauth2ApiReturnsTechnicalUserToken();
         irsApiSupport.irsApiTriggerJob();
-        irsApiSupport.irsApiReturnsJobDetails();
 
-        //WHEN
+        // WHEN/THEN
         given()
                 .contentType(ContentType.JSON)
                 .body(
@@ -63,14 +61,11 @@ class AssetAsBuiltControllerSyncIT extends IntegrationTestSpecification {
                 .then()
                 .statusCode(200);
 
-        //THEN
+
         eventually(() -> {
-            assetsSupport.assertAssetAsBuiltSize(15);
-            assetsSupport.assertHasRequiredIdentifiers();
-            assetsSupport.assertHasChildCount("urn:uuid:d387fa8e-603c-42bd-98c3-4d87fef8d2bb", 3);
+            irsApiSupport.verifyIrsApiTriggerJobCalledTimes(2);
             return true;
         });
-
     }
 
     @Test
@@ -97,7 +92,6 @@ class AssetAsBuiltControllerSyncIT extends IntegrationTestSpecification {
 
         //THEN
         eventually(() -> {
-            assetsSupport.assertAssetAsBuiltSize(15);
             irsApiSupport.verifyIrsApiTriggerJobCalledTimes(2);
             return true;
         });