+++ CHANGELOG.md
@@ -14,6 +14,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Added User Manual for Data import
 - Added github action to connect pull request with github issue
 - Added bpn validator to bpn edc mapping
+- Added new /irs/job/callback endpoint to handle irs job finished callback
 
 ### Changed
 - Fixed security findings
@@ -25,8 +26,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Fixed bug where applying a filter in one table erroneously affected sorting in the other table
 - Moved UpdateAssetRequest to tx-models module
 - updated Compatibility Matrix
+- Irs Jobs are now created with callback parameter set to new /irs/job/callback endpoint
 
 ### Removed
+- Shell descriptor entity with underlying logic
 
 ## [10.1.0 - 22.12.2023]
 ### Added